+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+<<<<<<< HEAD
 ### Added
 - Add health screener surveys [#2480](https://github.com/rokwire/illinois-app/issues/2480)
+=======
+### Changed
+- "Campus Guide Highlights" renamed to "Campus Safety Resources" [#2488](https://github.com/rokwire/illinois-app/issues/2488).
+>>>>>>> develop
 
 ## [4.2.10] - 2022-10-28
 ### Added